+++ include/nstd/gui.h
@@ -1,6 +1,7 @@
 #ifndef NSTD_GUI_H_INCLUDED
 #define NSTD_GUI_H_INCLUDED
 #include "gui/def.h"
-#include "gui/gui.h"
+#include "gui/display.h"
+#include "gui/window.h"
 #include "nstd.h"
 #endif